pyproject.toml
[project]
name = "rainymood-github-io"
version = "0.1.0"
description = ""
authors = ["Jan Meppe"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow (>=12.1.1,<13.0.0)",
    "numpy (>=2.4.4,<3.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)"
]

[tool.poetry]
packages = [{include = "rainymood"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
